add 3 problems
stick lenghts: median property, median has smallest overall distance beetween all numbers

missing coin: greedy summable stuff, suppose that you can compose all numbers from 0 up to a fixed number m - 1, then you check if you can compose the number m. if m is bigger than the array position we are checking (as there are no negative numbers) then we can compose all numbers from 0 up to m + arr[i] - 1, because i could just add arr[i] to the composition of m - 1, m - 2 ,.... 0 (if needed, i could also not add it and still have a composition for 1). else if its smaller that the array position we are checking now, we have no way to get that, because (as the array is ordered) i have no way to get that

restaurant: i'm just adding 1 everytime a client enters and removing it (in an efficient way, with priority queue) when it exists