Collecting numbers 2: The main idea of this second parts, that builds upon the main idea of the first part is that the number of rounds needed only depends on the position of the antecedent and posterior number. eg. if 7 is in position x, the number of rounds needed to calculate the contribution of 7 to the answer only depends on the position of 6 and 8, no need to focus on other things.
so, if, lets say, i have to swap the numbers at position a and b, and lets say these values are x and y
i know that the numbers x and y have only a limited contribution to the answer (0 min, 4 max), i calculate
the contribution at the current position, i swap them and calculate the contribution this time.
I know that this is all they contribute too, no need to focus on other things, so i just calculate the offset
and sum this to the answer before the swap. Thats all