add 1 problem
Collecting numbers 1:
Observation 1:
in one round, if i take x (lets say its at pos I0), i have to take x + 1, x + 2 and so on.
This is only possible if x + 1, x + 2 is in a position after I0, if i take x + 1 at position I1
i can take x + 2 iff I2 is after  I2, else i have to add a round in order to take it in this way.

So i can order the input array, and begin to take 1, 2 ... , n in order, counting if i can take with positions.